algunas actualizaciones de la documentacion y en el calculo de los stats

+++ src/main/java/com/EjercicioDetectorDeMutantes/EjercicioDetectorDeMutantes/dto/EstadisticasDTO.java
@@ -23,7 +23,7 @@ public class EstadisticasDTO {
 		if (cantidadDeDnaHumano == 0 || cantidadDeDnaMutantes == 0) {
 			proporcion =  "";
 		}else {
-			proporcion =  String.valueOf( cantidadDeDnaMutantes/cantidadDeDnaHumano);
+			proporcion =  String.valueOf((double) cantidadDeDnaMutantes/cantidadDeDnaHumano);
 		}
 	}
